requirements.txt
streamlit
speechrecognition
pyttsx3
textblob
pyaudio
together
python-dotenv
